Homework to lesson 2.1
