+++ src/LevelOne/lesson1_4/Main.java
@@ -1,0 +1,8 @@
+package LevelOne.lesson1_4;
+
+public class Main {
+    public static void main(String[] args) {
+        TikTakToe.start();
+    }
+
+}
