+++ src/LevelOne/lesson1_5/Main.java
@@ -1,0 +1,13 @@
+package LevelOne.lesson1_5;
+
+public class Main {
+    public static void main(String[] args) {
+
+        Worker.printInformation("Ivanov Ivan Ivanovich", "manager",
+                "ivan@gmail.com", "983498573", 50000, 50);
+
+        WorkersArray.findTheOldest();
+    }
+
+
+}
