+++ src/LevelOne/lesson1_6/Animal.java
@@ -1,0 +1,10 @@
+package LevelOne.lesson1_6;
+
+public abstract class Animal {
+
+    public void run(int lenght, String name){}
+
+    public void swim(int lenght, String name){}
+
+
+}
