+++ src/LevelOne/lesson1_8/Main.java
@@ -1,0 +1,7 @@
+package LevelOne.lesson1_8;
+
+public class Main {
+    public static void main(String[] args) {
+        new CalculatorFrame();
+    }
+}
