+++ src/LevelTwo/Lesson2_1/Entity/Entities.java
@@ -1,0 +1,9 @@
+package LevelTwo.Lesson2_1.Entity;
+
+public interface Entities {
+
+    boolean doRun(int distance);
+
+    boolean doJump(int height);
+
+}
