+++ src/LevelTwo/Lesson2_1/Lets/Lets.java
@@ -1,0 +1,9 @@
+package LevelTwo.Lesson2_1.Lets;
+
+import LevelTwo.Lesson2_1.Entity.Entities;
+
+public interface Lets {
+
+    boolean overcome (Entities entities);
+
+}
