--- src/lesson4/Main.java
@@ -1,8 +1,0 @@
-package lesson4;
-
-public class Main {
-    public static void main(String[] args) {
-        TikTakToe.start();
-    }
-
-}
