--- src/lesson5/Main.java
@@ -1,13 +1,0 @@
-package lesson5;
-
-public class Main {
-    public static void main(String[] args) {
-
-        Worker.printInformation("Ivanov Ivan Ivanovich", "manager",
-                "ivan@gmail.com", "983498573", 50000, 50);
-
-        WorkersArray.findTheOldest();
-    }
-
-
-}
