--- src/lesson6/Animal.java
@@ -1,10 +1,0 @@
-package lesson6;
-
-public abstract class Animal {
-
-    public void run(int lenght, String name){}
-
-    public void swim(int lenght, String name){}
-
-
-}
