--- src/lesson7/Cat.java
@@ -1,43 +1,0 @@
-package lesson7;
-
-public class Cat {
-
-    private String name;
-    private int appetite;
-    private boolean satiety = false;
-
-    public String getName() {
-        return name;
-    }
-
-    public void setName(String name) {
-        this.name = name;
-    }
-
-    public int getAppetite() {
-        return appetite;
-    }
-
-    public void setAppetite(int appetite) {
-        this.appetite = appetite;
-    }
-
-    public boolean isSatiety() {
-        return satiety;
-    }
-
-    public void setSatiety(boolean satiety) {
-        this.satiety = satiety;
-    }
-
-    public Cat(String name, int appetite, boolean satiety) {
-        this.name = name;
-        this.appetite = appetite;
-        this.satiety = satiety;
-    }
-
-    public void eat(Plate plate){
-        plate.decreaseFood(appetite);
-
-    }
-}
