--- src/lesson8/Main.java
@@ -1,7 +1,0 @@
-package lesson8;
-
-public class Main {
-    public static void main(String[] args) {
-        new CalculatorFrame();
-    }
-}
